Support for 64-bit iOS
NOT TESTED YET! But the iOS frameworks now include 64-bit binaries.
A side effect of this is that I had to bump the minimum supported iOS
version from 5.0 to 6.0, which for some reason is a prerequisite for
enabling 64-bit.
Fixes #130

+++ Demo-iOS/RootViewController.m
@@ -196,9 +196,10 @@
     NSUInteger numChecked = self.checkedDocuments.count;
     if (numChecked == 0)
         return;
-    NSString* message = [NSString stringWithFormat: @"Are you sure you want to remove the %u"
+    NSString* message = [NSString stringWithFormat: @"Are you sure you want to remove the %lu"
                                                      " checked-off item%@?",
-                                                     numChecked, (numChecked==1 ? @"" : @"s")];
+                                                     (unsigned long)numChecked,
+                                                     (numChecked==1 ? @"" : @"s")];
     UIAlertView* alert = [[UIAlertView alloc] initWithTitle: @"Remove Completed Items?"
                                                     message: message
                                                    delegate: self
